Add bare minimum palette functionality and testing
The command %palette may now give users a randomly selected palette from
colourpod.com, but there is as of yet no extra functionality around the
types of palettes users may choose, as was the case with Princess Luna.

This was implemented by doing a GET on the /random endpoint on the
Tumblr blog, which will fetch a random post and overwrite the request
message with that post's URL. This is a bit of a hack, however, as a
better solution would allow users to specify associated tags for some
palettes, as per original functionality.