Add wrapper for HttpClient and make singleton
This wrapper enables HTTP responses to be mocked out during unit
testing, and is a much cleaner way of doing so compared to mocking the
protected SendAsync method of the HttpMessageHandler class.

Although it creates a new `HttpClient` instance, this should be fine as
the wrapper class itself is set as a singleton in the ServiceCollection.